+++ NEWS.md
@@ -1,3 +1,19 @@
+# leaflet.esri 1.0.0
+
+## Leaflet.js 1.x
+
+- Uses package leaflet.js v1.3.1 by way of the `leaflet` R package
+- All javascript dependencies have been updated to the latest stable release.
+  - esri-leaflet-cluster@2.0.0 - https://github.com/Esri/esri-leaflet-cluster
+  - esri-leaflet-geocoder@2.2.9 - https://github.com/Esri/esri-leaflet-geocoder
+  - esri-leaflet-heatmap@2.0.0 - https://github.com/Esri/esri-leaflet-heatmap
+  - esri-leaflet-renderers@2.0.6 - https://github.com/Esri/esri-leaflet-renderers
+  - esri-leaflet@2.1.4 - https://github.com/Esri/esri-leaflet
+  - leaflet.heat@0.2.0 - https://github.com/Leaflet/Leaflet.heat
+  - leaflet.markercluster@1.3.0 - https://github.com/Leaflet/Leaflet.markercluster
+  - simpleheat@0.4.0 - https://github.com/mourner/simpleheat
+
+
 # leaflet.esri 0.2
 
 ## Major changes